+++ src/audiodevice.rs
@@ -512,6 +512,11 @@ pub fn new_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
     } else {
         conf.samplerate
     };
+    if let Some(cr) = conf.capture_samplerate {
+        if cr != conf.samplerate && conf.resampler.is_none() {
+            warn!("Resampling is disabled and capture_samplerate is different than samplerate, ignoring capture_samplerate.");
+        }
+    }
     let diff_rates = capture_samplerate != conf.samplerate;
     // Check for non-optimal resampling settings
     if !diff_rates && conf.resampler.is_some() && !conf.rate_adjust() {